Update product navbar
updated product navbar so the current product activates the toggleable
product list.  removed the 'products' button to make navigating a bit
clearer. had to add a dropdown arrow to include as the mortgage selected
is now a button instead of a dropdown. this makes it more obvious that
there are more products.