separated header, added reports page
changed the name of index.php to case.php and added a new index.php as
a work-in-progress for a reports page. Also separated out the main
nabber into header.html and using php include to bring the header in.
Adjusted the pages to allow for the fixed header.
Removed navbar from the main files to make it clearer to read. added
some graphical images to act as placeholders for reports